% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoCell.R
\name{IsoCell}
\alias{IsoCell}
\title{Isocell}
\usage{
IsoCell(gene_expr, isoform_expr, t = 0.1)
}
\arguments{
\item{gene_expr}{A gene-level expression data matrix, genes in rows and samples in columns.}

\item{isoform_expr}{An isoform-level expression data matrix, isoforms in rows and samples in columns.}

\item{t}{The retention ratio of isoforms, isoforms whose residual vector length is ranked in the top t percent are retained.}
}
\value{
P: The orthogonal projection matrix of genes.

len_residual_vector: The length of the residual vector for each normalized isoform vector.

selisoform: The isoforms IsoCell select for the final combining.

combined_data: The final combined data for clustering.
}
\description{
intergrating isoform-level expression through orthogonal projection
}
\examples{
data(demo)
result = IsoCell(gene_expr,isoform_expr)
}
\author{
Hongdong Li, lhdcsu@gmail.com, Central South University.
}
\keyword{orthogonal}
\keyword{projection}
